backend/requirements.txt
Django>=4.2
djangorestframework
pymongo
python-dotenv
python-dateutil
django-cors-headers
uvicorn
whitenoise